requirements.txt
jsonschema>=4.0.0
pytest>=8.0.0
python-dotenv>=1.0.0
gradio>=6.0.0,<7.0.0